[DOXIASITETOOLS-70] Html title of generated pages does not include project name
The problem was that the $title property was used for 2 different things in the velocity template: the title and the name of the last breadcrumb. They want different titles. The page title should include the project name, but the breadcrumb should not. I solved this by a adding a new property called $shortTitle that is now used by the breadcrumb generation.

git-svn-id: https://svn.apache.org/repos/asf/maven/doxia/doxia-sitetools/trunk@1309906 13f79535-47bb-0310-9956-ffa450edef68